Better system
Saves us lines, complexity and the need to calculate a super specific pose for each. Also, the apriltags are not equal on both sides.. if you want more information ask me, it's hard to explain right now. But, the transform from each apriltag to the position on the left, center and right are the same.