Add Driver Profiles (#19)
* Add DriverSettings class

* Remove stator current limiting

"Ramping" acts as the time to reach a throttle position from neutral. This means that stator current is not required. Some form of current limiting is required to avoid brown outs, so supply current configs are used.

* Add enum to convert names to DriverInfo

* Implement DriverInfo class to settings

Implement all the information stored in DriverInfo to the settings it needs to travel to.

* Remove unused swerve constants

* Remove unused Driver Profile updating stuff

* Fixed fatal progression error

* Static method because it's shorter

* Revert some settings and complete test run on 1/23