Reenable angle current limiting
* while also increasing the current limit

We need to figure out the best of both where the motors won't kill themselves, while also keeping the specificity and trying to overcome static friction. Might look into better angle motor PID as it is the only slightly tuned, and it was tuned in the air so it does not account for friction.